chore: remove dead TRANSCRIBE_INSTRUCTION command, fold stray modelHandlers tests
Delete the TRANSCRIBE_INSTRUCTION command chain (ipc constant, inbound
schema entry, MainViewMessageHandler registry binding, InstructionManager
stub handler) — the webview no longer sends this command; its only
behavior was a "feature moved" notice with zero senders repo-wide.

Merge the off-convention src/test-kernel/modelHandlers/ directory into
the canonical src/test-kernel/agent/modelHandlers/ mirror of src/agent/modelHandlers/,
folding ModelHandlerGoogleGenAI shouldContinue/extractToolUse/createResponse
coverage and toGoogleTools coverage into their canonical suites, and
dropping the toOpenAITools/toOpenAIResponseTools cases that already exist
in ToolConversion.vitest.ts.
