+++ src/test-kernel/agent/modelHandlers/ToolConversion.vitest.ts
@@ -1,8 +1,11 @@
+import { strict as assert } from 'node:assert';
+
 import { describe, expect, it } from 'vitest';
 import { z } from 'zod';
 
 import {
   toAnthropicTools,
+  toGoogleTools,
   toOpenAITools,
   toOpenAIResponseTools,
 } from '@agent/modelHandlers/toolConversion';
@@ -21,6 +24,7 @@ import { CrossrefSearchTool } from '@tools/citation/CrossrefSearchTool';
 import { TexcountTool } from '@tools/texcount/TexcountTool';
 import { WebFetchTool } from '@tools/web/WebFetchTool';
 import { WebSearchTool } from '@tools/web/WebSearchTool';
+import type { Tool as GeminiTool } from '@google/genai';
 import type { ChatCompletionTool } from 'openai/resources/chat/completions';
 import type { FunctionTool } from 'openai/resources/responses/responses';
 
@@ -410,5 +414,260 @@ describe('Crossref provider schema compatibility', () => {
         unknown: true,
       }).success,
     ).toBe(false);
+  });
+});
+
+describe('toOpenAITools additional coverage', () => {
+  it('leaves Chat Completions function tools non-strict', () => {
+    const defs: ToolDefinition[] = [
+      {
+        name: 'delegate_workflow',
+        description: 'Delegate workflow',
+        parameters: {
+          type: 'object',
+          properties: { instruction: { type: 'string' } },
+          required: ['instruction'],
+          additionalProperties: false,
+        },
+      },
+    ];
+
+    const tools = toOpenAITools(defs);
+    assert.equal(tools.length, 1);
+    assert.equal(tools[0].type, 'function');
+    assert.equal(tools[0].function.name, 'delegate_workflow');
+    assert.equal(tools[0].function.strict, undefined);
+  });
+});
+
+describe('toOpenAIResponseTools', () => {
+  it('converts tool definitions to Response API format', () => {
+    const defs: ToolDefinition[] = [
+      {
+        name: 'echo',
+        description: 'Echo value',
+        parameters: {
+          type: 'object',
+          properties: { value: { type: 'string' } },
+          required: ['value'],
+          additionalProperties: false,
+        },
+      },
+    ];
+
+    const tools = toOpenAIResponseTools(defs);
+    assert.equal(tools.length, 1);
+    const tool = tools[0] as FunctionTool;
+    assert.equal(tool.type, 'function');
+    assert.equal(tool.name, 'echo');
+    assert.deepEqual(tool.parameters, defs[0].parameters);
+  });
+
+  // web_search collapses to the native WebSearchTool when (and only when) the
+  // model supports it, regardless of whether function calling is also enabled.
+  it.each([
+    {
+      label:
+        'converts web_search to native WebSearchTool when supportsNativeWebSearch is true',
+      defs: [{ name: 'web_search', description: 'Search the web' }],
+      options: { supportsNativeWebSearch: true },
+    },
+    {
+      label:
+        'keeps native web_search when supportsFunctionCalling is false and supportsNativeWebSearch is true',
+      defs: [
+        { name: 'web_search', description: 'Search the web' },
+        { name: 'read_file', description: 'Read a file' },
+      ],
+      options: {
+        supportsFunctionCalling: false,
+        supportsNativeWebSearch: true,
+      },
+    },
+  ])('$label', ({ defs, options }) => {
+    const tools = toOpenAIResponseTools(defs as ToolDefinition[], options);
+    assert.equal(tools.length, 1);
+    assert.equal(tools[0].type, 'web_search');
+  });
+
+  // Any tool stays a plain function tool when native web search is off, which is
+  // also the default for both options.
+  it.each([
+    {
+      label:
+        'keeps web_search as function tool when supportsNativeWebSearch is false',
+      def: { name: 'web_search', description: 'Search the web' },
+      options: { supportsNativeWebSearch: false },
+    },
+    {
+      label: 'defaults supportsNativeWebSearch to false',
+      def: { name: 'web_search', description: 'Search the web' },
+      options: undefined,
+    },
+    {
+      label: 'defaults supportsFunctionCalling to true',
+      def: { name: 'custom_tool', description: 'A custom function tool' },
+      options: undefined,
+    },
+  ])('$label', ({ def, options }) => {
+    const defs: ToolDefinition[] = [def];
+    // The default cases call with no options argument, exactly as before.
+    const tools = options
+      ? toOpenAIResponseTools(defs, options)
+      : toOpenAIResponseTools(defs);
+    assert.equal(tools.length, 1);
+    const tool = tools[0] as FunctionTool;
+    assert.equal(tool.type, 'function');
+    assert.equal(tool.name, def.name);
+  });
+
+  it('filters out function tools when supportsFunctionCalling is false', () => {
+    const defs: ToolDefinition[] = [
+      {
+        name: 'read_file',
+        description: 'Read a file',
+        parameters: {
+          type: 'object',
+          properties: { path: { type: 'string' } },
+          required: ['path'],
+        },
+      },
+      {
+        name: 'write_file',
+        description: 'Write a file',
+      },
+    ];
+
+    // Deep research models don't support function calling, so both drop out.
+    const tools = toOpenAIResponseTools(defs, {
+      supportsFunctionCalling: false,
+    });
+    assert.equal(tools.length, 0);
+  });
+});
+
+describe('toGoogleTools', () => {
+  // NOTE: Native googleSearch is disabled because Google's regular content
+  // generation API does NOT support combining googleSearch with functionDeclarations.
+  // This is a Live API only feature. All tools are converted to function declarations.
+
+  it('returns empty array for empty input', () => {
+    const tools = toGoogleTools([]);
+    assert.deepEqual(tools, []);
+  });
+
+  // Every tool (including web_search) becomes a function declaration wrapped in a
+  // single Tool object; native googleSearch is never emitted.
+  it.each([
+    {
+      label: 'converts function declarations to single tool object',
+      defs: [
+        {
+          name: 'read_file',
+          description: 'Read a file',
+          parameters: {
+            type: 'object',
+            properties: { path: { type: 'string' } },
+            required: ['path'],
+          },
+        },
+      ],
+      names: ['read_file'],
+    },
+    {
+      label:
+        'converts web_search to function declaration (native googleSearch disabled)',
+      defs: [{ name: 'web_search', description: 'Search the web' }],
+      names: ['web_search'],
+    },
+    {
+      label: 'converts all tools to function declarations',
+      defs: [
+        { name: 'web_search', description: 'Search the web' },
+        {
+          name: 'read_file',
+          description: 'Read a file',
+          parameters: {
+            type: 'object',
+            properties: { path: { type: 'string' } },
+            required: ['path'],
+          },
+        },
+        {
+          name: 'write_file',
+          description: 'Write a file',
+          parameters: {
+            type: 'object',
+            properties: {
+              path: { type: 'string' },
+              content: { type: 'string' },
+            },
+            required: ['path', 'content'],
+          },
+        },
+      ],
+      names: ['web_search', 'read_file', 'write_file'],
+    },
+  ])('$label', ({ defs, names }) => {
+    const tools = toGoogleTools(defs as ToolDefinition[]);
+    assert.equal(tools.length, 1);
+    const tool = tools[0] as GeminiTool;
+    assert.equal(tool.googleSearch, undefined);
+    assert.ok(tool.functionDeclarations);
+    assert.deepEqual(
+      tool.functionDeclarations?.map((fd) => fd.name),
+      names,
+    );
+  });
+
+  it('flattens top-level discriminated unions into a single object schema', () => {
+    // Gemini rejects function parameter schemas whose root is oneOf/anyOf/allOf
+    // (HTTP 400: schema must have type 'object' and not contain 'oneOf'/'anyOf'/'allOf').
+    // Discriminated unions like the inquiry tool must flatten to a single object.
+    const defs: ToolDefinition[] = [
+      {
+        name: 'inquiry',
+        description: 'Ask, read, or list inquiry threads',
+        zodSchema: z.discriminatedUnion('command', [
+          z.object({
+            command: z.literal('ask'),
+            question: z.string(),
+            thread_id: z.string().nullish(),
+          }),
+          z.object({
+            command: z.literal('read'),
+            thread_id: z.string(),
+          }),
+          z.object({
+            command: z.literal('list'),
+            status: z.enum(['open', 'answered']).default('open'),
+          }),
+        ]),
+      },
+    ];
+
+    const tools = toGoogleTools(defs);
+    const tool = tools[0] as GeminiTool;
+    const params = tool.functionDeclarations?.[0]
+      .parametersJsonSchema as Record<string, unknown>;
+
+    assert.equal(params.type, 'object');
+    assert.equal(params.oneOf, undefined);
+    assert.equal(params.anyOf, undefined);
+    assert.equal(params.allOf, undefined);
+
+    const properties = params.properties as Record<
+      string,
+      Record<string, unknown>
+    >;
+    // Discriminator collapses literal branches into an enum.
+    assert.deepEqual(properties.command.enum, ['ask', 'read', 'list']);
+    assert.equal(properties.command.type, 'string');
+    // Branch-specific props are merged in.
+    assert.ok(properties.question);
+    assert.ok(properties.thread_id);
+    assert.ok(properties.status);
+    // Only command is required by every branch.
+    assert.deepEqual(params.required, ['command']);
   });
 });
